feat: code refactoring and structural improvement
